+++ R/report.R
@@ -7,6 +7,12 @@
 #'
 #' @param df A data frame containing multiple time series in long format. See
 #'   Details.
+#' @param file String specifying the desired file name (and path) to save the
+#'   report to. The file name should include the extension ".html". If only a
+#'   file name is supplied, the report will be saved in the current working
+#'   directory. If a path is supplied, the directory should already exist. Any
+#'   existing file of the same name will be overwritten unless
+#'   `add_timestamp` is set to `TRUE`.
 #' @param inputspec [`inputspec()`] object specifying which columns in the
 #'   supplied `df` represent the "timepoint", "item", "value"  and (optionally)
 #'   "tab" for the time series. If a "tab" column is specified, a separate tab
@@ -14,19 +20,17 @@
 #' @param outputspec `outputspec` object specifying the desired format of the
 #'   html table(s). If not supplied, default values will be used.
 #' @param alertspec [`alertspec()`] object specifying conditions to test and
-#'   display
+#'   display.
 #' @param report_title Title to appear on the report.
 #' @param dataset_description Short description of the dataset being shown. This
 #'   will appear on the report.
-#' @param save_directory String specifying directory in which to save the
-#'   report. Default is current directory.
-#' @param save_filename String specifying filename for the report, excluding any
-#'   file extension. If no filename is supplied, one will be automatically
-#'   generated with the format `mantis_report_YYMMDD_HHMMSS`.
+#' @param add_timestamp Append a timestamp to the end of the filename with
+#'   format `_YYMMDD_HHMMSS`. This can be used to keep multiple versions of the
+#'   same report. Default = `FALSE`.
 #' @param show_progress Print progress to console. Default = `TRUE`.
 #' @param ... Further parameters to be passed to `rmarkdown::render()`. Cannot
 #'   include any of `input`, `output_dir`, `output_file`, `params`, `quiet`.
-#' @return A string containing the name and path of the saved report.
+#' @return A string containing the name and full path of the saved report.
 #'
 #' @details The supplied data frame should contain multiple time series in long
 #'   format, i.e.:
@@ -44,10 +48,11 @@
 #'   The `inputspec` parameter maps the data frame columns to the above.
 #' @examples
 #' \donttest{
-#' # create an interactive report in the current directory,
+#' # create an interactive report in the temp directory,
 #' # with one tab per Location
 #' filename <- mantis_report(
 #'   df = example_prescription_numbers,
+#'   file = file.path(tempdir(), "example_prescription_numbers_interactive.html"),
 #'   inputspec = inputspec(
 #'     timepoint_col = "PrescriptionDate",
 #'     item_cols = c("Location", "Antibiotic", "Spectrum"),
@@ -58,17 +63,18 @@
 #'   outputspec = outputspec_interactive(),
 #'   report_title = "Daily antibiotic prescribing",
 #'   dataset_description = "Antibiotic prescriptions by site",
-#'   save_directory = ".",
-#'   save_filename = "example_prescription_numbers_interactive",
 #'   show_progress = TRUE
 #' )
-#' \dontshow{file.remove("./example_prescription_numbers_interactive.html")}
+#'
+#' filename
+#' \dontshow{unlink(filename)}
 #' }
 #'
 #' \donttest{
-#' # create an interactive report in the current directory, with alerting rules
+#' # create an interactive report in the temp directory, with alerting rules
 #' filename <- mantis_report(
 #'   df = example_prescription_numbers,
+#'   file = file.path(tempdir(), "example_prescription_numbers_interactive.html"),
 #'   inputspec = inputspec(
 #'     timepoint_col = "PrescriptionDate",
 #'     item_cols = c("Location", "Antibiotic", "Spectrum"),
@@ -86,11 +92,11 @@
 #'   ),
 #'   report_title = "Daily antibiotic prescribing",
 #'   dataset_description = "Antibiotic prescriptions by site",
-#'   save_directory = ".",
-#'   save_filename = "example_prescription_numbers_interactive",
 #'   show_progress = TRUE
 #' )
-#' \dontshow{file.remove("./example_prescription_numbers_interactive.html")}
+#'
+#' filename
+#' \dontshow{unlink(filename)}
 #' }
 #'
 #' @seealso [inputspec()], [outputspec_interactive()],
@@ -99,13 +105,13 @@
 #' @export
 mantis_report <- function(
   df,
+  file,
   inputspec,
   outputspec = NULL,
   alertspec = NULL,
   report_title = "mantis report",
   dataset_description = "",
-  save_directory = ".",
-  save_filename = NULL,
+  add_timestamp = FALSE,
   show_progress = TRUE,
   ...
 ) {
@@ -113,39 +119,37 @@ mantis_report <- function(
   validate_params_type(
     match.call(),
     df = df,
+    file = file,
     inputspec = inputspec,
     outputspec = outputspec,
     alertspec = alertspec,
     report_title = report_title,
     dataset_description = dataset_description,
-    save_directory = save_directory,
-    save_filename = save_filename,
+    add_timestamp = add_timestamp,
     show_progress = show_progress
   )
 
   validate_df_to_inputspec(df, inputspec)
   validate_alert_rules_to_inputspec(alertspec$alert_rules, inputspec)
 
-  timestamp_string <- format(Sys.time(), "%Y%m%d%_%H%M%S")
+  timestamp_string <- format(Sys.time(), "_%Y%m%d%_%H%M%S")
 
   if (is.null(outputspec)) {
     outputspec <- outputspec_interactive()
   }
 
-  if (is.null(save_filename)) {
-    save_filename <-
-      paste0("mantis_", timestamp_string)
-  }
+  file_fullname <- format_filename(
+    file,
+    append_string = ifelse(add_timestamp, timestamp_string, "")
+  )
 
-  file_and_path <- file.path(save_directory, paste0(save_filename, ".html"))
-
-  # temporarily copy rmd file from package library into save_directory so that
+  # temporarily copy rmd file from package library into output directory so that
   # intermediate files also get created there.
   # NOTE: explicitly setting intermediates_dir in rmarkdown::render() to
-  # save_directory or tempdir() causes duplicate chunk label errors when package
+  # output directory or tempdir() causes duplicate chunk label errors when package
   # is run from inside an rmd/qmd
   temp_dirname <-
-    file.path(save_directory, paste0("mantis_temp_", timestamp_string))
+    file.path(dirname(file_fullname), paste0("mantis_temp", timestamp_string))
   dir.create(temp_dirname)
   file.copy(
     from = system.file(
@@ -160,8 +164,7 @@ mantis_report <- function(
 
   rmarkdown::render(
     input = file.path(temp_dirname, "report-html.Rmd"),
-    output_file = paste0(save_filename, ".html"),
-    output_dir = save_directory,
+    output_file = file_fullname,
     params = list(
       df = df,
       inputspec = inputspec,
@@ -177,5 +180,28 @@ mantis_report <- function(
   # remove temporary directory created earlier
   unlink(temp_dirname, recursive = TRUE)
 
-  file_and_path
+  file_fullname
+}
+
+
+
+#' Format the supplied file name and path if necessary
+#'
+#' This assumes the file path and name has already passed validation.
+#'
+#' @param file string containing file name (and optionally path), with the
+#'   extension .html
+#' @param append_string append the string to file name, before the extension.
+#'
+#' @returns A string containing the formatted file name and canonical path
+#' @noRd
+format_filename <- function(
+  file,
+  append_string = ""
+) {
+  path <- normalizePath(dirname(file))
+  filename_sans_ext <- tools::file_path_sans_ext(basename(file))
+  ext <- tools::file_ext(file)
+
+  file.path(path, paste0(filename_sans_ext, append_string, ".", ext))
 }