+++ tests/testthat/test-report.R
@@ -12,14 +12,13 @@ test_that("mantis_report() creates single-tab interactive report and returns pat
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = "item",
         value_col = "value"
       ),
       outputspec = outputspec_interactive(),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
 
@@ -44,6 +43,7 @@ test_that("mantis_report() creates multi-tab interactive report and returns path
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = c("item", "tab"),
@@ -51,8 +51,6 @@ test_that("mantis_report() creates multi-tab interactive report and returns path
         tab_col = "tab"
       ),
       outputspec = outputspec_interactive(),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
 
@@ -78,14 +76,13 @@ test_that("mantis_report() creates heatmap report and returns path successfully"
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = "item",
         value_col = "value"
       ),
       outputspec = outputspec_static_heatmap(),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
   expect_type(reportpath, "character")
@@ -95,6 +92,7 @@ test_that("mantis_report() creates heatmap report and returns path successfully"
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = c("item", "tab"),
@@ -102,8 +100,6 @@ test_that("mantis_report() creates heatmap report and returns path successfully"
         tab_col = "tab"
       ),
       outputspec = outputspec_static_heatmap(),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
   expect_type(reportpath, "character")
@@ -126,14 +122,13 @@ test_that("mantis_report() creates multipanel report and returns path successful
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = "item",
         value_col = "value"
       ),
       outputspec = outputspec_static_multipanel(),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
   expect_type(reportpath, "character")
@@ -143,6 +138,7 @@ test_that("mantis_report() creates multipanel report and returns path successful
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = c("item", "tab"),
@@ -150,8 +146,6 @@ test_that("mantis_report() creates multipanel report and returns path successful
         tab_col = "tab"
       ),
       outputspec = outputspec_static_multipanel(),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
   expect_type(reportpath, "character")
@@ -173,6 +167,7 @@ test_that("mantis_report() creates interactive report with alerts successfully",
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = c("item", "tab"),
@@ -194,8 +189,6 @@ test_that("mantis_report() creates interactive report with alerts successfully",
           )
         )
       ),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
 
@@ -218,14 +211,13 @@ test_that("mantis_report() creates interactive report even if supplied an empty 
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = "item",
         value_col = "value"
       ),
       outputspec = outputspec_interactive(),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
   expect_true(file.remove(reportpath))
@@ -234,6 +226,7 @@ test_that("mantis_report() creates interactive report even if supplied an empty 
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = c("item", "tab"),
@@ -241,8 +234,6 @@ test_that("mantis_report() creates interactive report even if supplied an empty 
         tab_col = "tab"
       ),
       outputspec = outputspec_interactive(),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
   expect_true(file.remove(reportpath))
@@ -259,6 +250,7 @@ test_that("mantis_report() creates interactive report with alerts even if suppli
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = c("item", "tab"),
@@ -280,8 +272,6 @@ test_that("mantis_report() creates interactive report with alerts even if suppli
           )
         )
       ),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
 
@@ -301,14 +291,13 @@ test_that("mantis_report() creates heatmap report even if supplied an empty df",
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = "item",
         value_col = "value"
       ),
       outputspec = outputspec_static_heatmap(),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
   expect_true(file.remove(reportpath))
@@ -317,6 +306,7 @@ test_that("mantis_report() creates heatmap report even if supplied an empty df",
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = c("item", "tab"),
@@ -324,8 +314,6 @@ test_that("mantis_report() creates heatmap report even if supplied an empty df",
         tab_col = "tab"
       ),
       outputspec = outputspec_static_heatmap(),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
   expect_true(file.remove(reportpath))
@@ -343,14 +331,13 @@ test_that("mantis_report() creates multipanel report even if supplied an empty d
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = "item",
         value_col = "value"
       ),
       outputspec = outputspec_static_multipanel(),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
   expect_true(file.remove(reportpath))
@@ -359,6 +346,7 @@ test_that("mantis_report() creates multipanel report even if supplied an empty d
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = c("item", "tab"),
@@ -366,8 +354,6 @@ test_that("mantis_report() creates multipanel report even if supplied an empty d
         tab_col = "tab"
       ),
       outputspec = outputspec_static_multipanel(),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
   expect_true(file.remove(reportpath))
@@ -388,14 +374,13 @@ test_that("mantis_report() creates multi-item_cols interactive report and return
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "timepoint",
         item_cols = c("item", "tab"),
         value_col = "value"
       ),
       outputspec = outputspec_interactive(item_labels = c("item" = "My item")),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
 
@@ -422,6 +407,7 @@ test_that("mantis_report() creates interactive report with spaces in item_cols n
   reportpath <-
     mantis_report(
       df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
       inputspec = inputspec(
         timepoint_col = "m timepoint",
         item_cols = c("m item", "m tab"),
@@ -443,12 +429,136 @@ test_that("mantis_report() creates interactive report with spaces in item_cols n
           )
         )
       ),
-      save_directory = tempdir(),
-      save_filename = "mantis_testthatreport",
       show_progress = FALSE
     )
 
   expect_type(reportpath, "character")
+
+  # clean up
+  expect_true(file.remove(reportpath))
+})
+
+
+test_that("format_filename() appends explict timestamp correctly when a path and name is supplied", {
+  timestamp_string <- format(Sys.time(), "_%Y%m%d%_%H%M%S")
+  orig_file <- file.path(test_path(), "filename.html")
+  new_file <- file.path(test_path(), paste0("filename", timestamp_string, ".html"))
+
+  # appends to filename
+  expect_equal(
+    basename(format_filename(file = orig_file,
+                             append_string = timestamp_string)),
+    basename(new_file)
+  )
+  # but does not change path
+  expect_equal(
+    normalizePath(dirname(format_filename(file = orig_file,
+                             append_string = timestamp_string))),
+    normalizePath(dirname(new_file))
+  )
+})
+
+
+test_that("format_filename() appends explict timestamp correctly when only a name is supplied", {
+  timestamp_string <- format(Sys.time(), "_%Y%m%d%_%H%M%S")
+  orig_file <- "filename.html"
+  new_file <- paste0("filename", timestamp_string, ".html")
+
+  # appends to filename
+  expect_equal(
+    basename(format_filename(file = orig_file,
+                             append_string = timestamp_string)),
+    basename(new_file)
+  )
+  # but does not change path
+  expect_equal(
+    normalizePath(dirname(format_filename(file = orig_file,
+                                          append_string = timestamp_string))),
+    normalizePath(dirname(new_file))
+  )
+})
+
+
+test_that("mantis_report() appends timestamp to filename appropriately", {
+  df <- data.frame(
+    timepoint = rep(
+      seq(as.Date("2022-01-01"), as.Date("2022-01-10"), by = "days"),
+      2
+    ),
+    item = c(rep("a", 10), rep("b", 10)),
+    value = rep(c(3, -3), 10),
+    stringsAsFactors = FALSE
+  )
+
+  # include timestamp
+  reportpath <-
+    mantis_report(
+      df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
+      inputspec = inputspec(
+        timepoint_col = "timepoint",
+        item_cols = "item",
+        value_col = "value"
+      ),
+      outputspec = outputspec_interactive(),
+      add_timestamp = TRUE,
+      show_progress = FALSE
+    )
+
+  expect_true(grepl("^mantis_testthatreport_[0-9]{8}_[0-9]{6}\\.html$", basename(reportpath)))
+
+  # clean up
+  expect_true(file.remove(reportpath))
+
+  # exclude timestamp
+  reportpath <-
+    mantis_report(
+      df,
+      file = file.path(tempdir(), "mantis_testthatreport.html"),
+      inputspec = inputspec(
+        timepoint_col = "timepoint",
+        item_cols = "item",
+        value_col = "value"
+      ),
+      outputspec = outputspec_interactive(),
+      add_timestamp = FALSE,
+      show_progress = FALSE
+    )
+
+  expect_equal(basename(reportpath), "mantis_testthatreport.html")
+
+  # clean up
+  expect_true(file.remove(reportpath))
+})
+
+
+test_that("mantis_report() creates report in wd when no path supplied", {
+  df <- data.frame(
+    timepoint = rep(
+      seq(as.Date("2022-01-01"), as.Date("2022-01-10"), by = "days"),
+      2
+    ),
+    item = c(rep("a", 10), rep("b", 10)),
+    value = rep(c(3, -3), 10),
+    stringsAsFactors = FALSE
+  )
+
+  withr::with_dir(
+    tempdir(),
+    reportpath <-
+      mantis_report(
+        df,
+        file = "mantis_testthatreport.html",
+        inputspec = inputspec(
+          timepoint_col = "timepoint",
+          item_cols = "item",
+          value_col = "value"
+        ),
+        outputspec = outputspec_interactive(),
+        add_timestamp = TRUE,
+        show_progress = FALSE
+      )
+  )
 
   # clean up
   expect_true(file.remove(reportpath))